PDF_Query_using_langchain_google_genai/requirements.txt
ipykernel
python-dotenv
PyPDF2
pypdf 
langchain
langchain_community
langchain-google-genai
streamlit
faiss-cpu